[LAS-12] refactor: member 업데이트 코드 리팩토링
가독성 저해 요소인 이중 if문 분리 작업

+++ api-server/src/main/java/com/lastone/apiserver/service/member/MemberServiceImpl.java
@@ -10,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
@@ -33,12 +32,12 @@ public class MemberServiceImpl implements MemberService {
     @Transactional(rollbackFor = Exception.class)
     public void update(Member member, MemberUpdateDto memberUpdateDto, MultipartFile profileImg) throws IOException {
         isDuplicatedNickname(memberUpdateDto.getNickname(), member.getNickname());
-
-        if (!ObjectUtils.isEmpty(profileImg)) {
-            if (StringUtils.hasText(member.getProfileUrl())) {
-                s3Service.delete(member.getProfileUrl());
-            }
-            memberUpdateDto.setProfileUrl(s3Service.upload(profileImg));
+        if (!profileImg.isEmpty() && StringUtils.hasText(member.getProfileUrl())) {
+            s3Service.delete(member.getProfileUrl());
+        }
+        if (!profileImg.isEmpty()) {
+            String profileUrl = s3Service.upload(profileImg);
+            memberUpdateDto.setProfileUrl(profileUrl);
         }
         member.update(memberUpdateDto);
     }
